# cedar-user-server-application/src/main/java/org/metadatacenter/cedar/user/resources/UsersResource.java
package org.metadatacenter.cedar.user.resources;

import com.codahale.metrics.annotation.Timed;
import com.fasterxml.jackson.databind.JsonNode;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.parameters.RequestBody;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import org.keycloak.admin.client.Keycloak;
import org.keycloak.admin.client.resource.UserResource;
import org.keycloak.representations.idm.FederatedIdentityRepresentation;
import org.keycloak.representations.idm.UserRepresentation;
import org.metadatacenter.config.CedarConfig;
import org.metadatacenter.error.CedarErrorKey;
import org.metadatacenter.exception.CedarDependencyUnavailableException;
import org.metadatacenter.exception.CedarException;
import org.metadatacenter.id.CedarUserId;
import org.metadatacenter.rest.context.CedarRequestContext;
import org.metadatacenter.server.result.BackendCallResult;
import org.metadatacenter.server.security.KeycloakUtilInfo;
import org.metadatacenter.server.security.KeycloakUtils;
import org.metadatacenter.server.security.model.auth.CedarPermission;
import org.metadatacenter.server.security.model.user.CedarUser;
import org.metadatacenter.server.security.model.user.CedarUserApiKey;
import org.metadatacenter.server.service.UserService;
import org.metadatacenter.util.CedarUserNameUtil;
import org.metadatacenter.util.http.CedarResponse;
import org.metadatacenter.util.json.JsonMapper;
import org.metadatacenter.util.mongo.MongoUtils;

import jakarta.ws.rs.DELETE;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.NotFoundException;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.PUT;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.ProcessingException;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;
import java.security.SecureRandom;
import java.net.URI;
import java.time.LocalDate;
import java.time.LocalDateTime;
import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import java.util.UUID;

import static org.metadatacenter.constant.CedarPathParameters.PP_ID;
import static org.metadatacenter.rest.assertion.GenericAssertions.LoggedIn;

@Path("/users")
@Produces(MediaType.APPLICATION_JSON)
@Tag(name = "Users")
@SecurityRequirement(name = "api_key")
public class UsersResource extends AbstractUserServerResource {

  private static final int MAX_API_KEYS = 20;

  private static UserService userService;
  private KeycloakUtilInfo kcInfo;
  private Keycloak kc;

  public UsersResource(CedarConfig cedarConfig) {
    super(cedarConfig);
    kcInfo = KeycloakUtils.initKeycloak(cedarConfig);
    kc = KeycloakUtils.buildKeycloak(kcInfo);
  }

  public static void injectUserService(UserService us) {
    userService = us;
  }

  @GET
  @Timed
  @Path("/{id}")
  @Operation(summary = "Get the current user's profile",
      description = "Returns the authenticated user's complete CEDAR profile. A user cannot read another user's profile.")
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The current user's profile",
          content = @Content(schema = @Schema(implementation = CedarUser.class))),
      @ApiResponse(responseCode = "401", description = "Unauthorized"),
      @ApiResponse(responseCode = "403", description = "Another user's profile was requested"),
      @ApiResponse(responseCode = "500", description = "Internal server error")
  })
  public Response findOwnUser(@Parameter(description = "CEDAR user UUID", required = true)
                              @PathParam(PP_ID) String uuid) throws CedarException {
    CedarRequestContext c = buildRequestContext();

    c.must(c.user()).be(LoggedIn);

    CedarUser currentUser = c.getCedarUser();

    String id = linkedDataUtil.getUserId(uuid);

    if (!id.equals(currentUser.getId())) {
      return CedarResponse.forbidden()
          .id(id)
          .errorKey(CedarErrorKey.READ_OTHER_PROFILE_FORBIDDEN)
          .errorMessage("You are not allowed to read other user's profile!")
          .parameter("currentUserId", currentUser.getId())
          .build();
    }

    JsonNode user = JsonMapper.MAPPER.valueToTree(currentUser);
    // Remove autogenerated _id field to avoid exposing it
    MongoUtils.removeIdField(user);

    return Response.ok().entity(user).build();
  }

  @GET
  @Timed
  @Path("/{id}/summary")
  @Operation(summary = "Get a user summary",
      description = "Returns display and federated-identity information. Self access and authorized administrator access only.")
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "User summary",
          content = @Content(schema = @Schema(type = "object"))),
      @ApiResponse(responseCode = "401", description = "Unauthorized"),
      @ApiResponse(responseCode = "403", description = "Forbidden"),
      @ApiResponse(responseCode = "404", description = "User not found"),
      @ApiResponse(responseCode = "503", description = "Keycloak is unavailable"),
      @ApiResponse(responseCode = "500", description = "Internal server error")
  })
  public Response findUserSummary(@Parameter(description = "CEDAR user UUID", required = true)
                                  @PathParam(PP_ID) String uuid) throws CedarException {
    CedarRequestContext c = buildRequestContext();

    c.must(c.user()).be(LoggedIn);

    String id = linkedDataUtil.getUserId(uuid);
    CedarUser currentUser = c.getCedarUser();

    // Self, or a user administrator. USER_READ belongs to the userAdministrator role, which the
    // built-in admin holds and an ordinary user does not, so this admits the provenance display-name
    // lookup UserSummaryCache makes as the admin while refusing one user's summary to another. The
    // summary carries each federated identity provider together with the account id held there,
    // which is the user's identifier at Google or ORCID.
    if (!id.equals(currentUser.getId()) && !currentUser.has(CedarPermission.USER_READ)) {
      return CedarResponse.forbidden()
          .id(id)
          .errorKey(CedarErrorKey.READ_OTHER_PROFILE_FORBIDDEN)
          .errorMessage("You are not allowed to read other user's profile!")
          .parameter("currentUserId", currentUser.getId())
          .build();
    }

    CedarUserId uid = CedarUserId.build(id);

    CedarUser lookupUser = userService.findUser(uid);
    // An id CEDAR does not hold produced a 200 carrying a null screen name, since getDisplayName
    // answers null for a null user.
    if (lookupUser == null) {
      return CedarResponse.notFound()
          .id(id)
          .errorKey(CedarErrorKey.USER_NOT_FOUND)
          .errorMessage("The user can not be found by id!")
          .build();
    }

    UserRepresentation keyCloakUserRepresentation;
    try {
      UserResource userResource = kc.realm(kcInfo.getKeycloakRealmName()).users().get(uuid);
      keyCloakUserRepresentation = userResource.toRepresentation();
    } catch (NotFoundException e) {
      // Keycloak reports an unknown user by throwing, which left through the generic mapper as a
      // bare 404 carrying no CEDAR error body.
      return CedarResponse.notFound()
          .id(id)
          .errorKey(CedarErrorKey.USER_NOT_FOUND)
          .errorMessage("The user can not be found by id!")
          .build();
    } catch (ProcessingException e) {
      throw new CedarDependencyUnavailableException("Keycloak is unavailable", e);
    }

    Map<String, Object> summary = new HashMap<>();
    summary.put("userId", id);
    summary.put("screenName", CedarUserNameUtil.getDisplayName(cedarConfig, lookupUser));
    // Keycloak account creation time (epoch millis) - used as "Member since" on the profile page.
    summary.put("createdTimestamp", keyCloakUserRepresentation.getCreatedTimestamp());

    if (keyCloakUserRepresentation.getFederatedIdentities() != null) {
      List<Object> authenticationProviders = new ArrayList<>();

      for (FederatedIdentityRepresentation federatedIdentityRepresentation : keyCloakUserRepresentation.getFederatedIdentities()) {
        Map<String, String> authenticationProviderUserPair = new HashMap<>();
        String authenticationProvider = federatedIdentityRepresentation.getIdentityProvider();
        String userID = federatedIdentityRepresentation.getUserId();

        authenticationProviderUserPair.put("name", authenticationProvider);
        authenticationProviderUserPair.put("id", userID);
        authenticationProviders.add(authenticationProviderUserPair);
      }
      summary.put("authenticationProvider", authenticationProviders);
    }

    return Response.ok().entity(summary).build();
  }

  @PUT
  @Timed
  @Path("/{id}")
  @Operation(summary = "Patch the current user's UI preferences", description = "Applies the supplied dotted-path "
      + "preference changes to the current stored profile under a Neo4j write lock. Only uiPreferences fields are "
      + "accepted, and the write cannot overwrite API keys, roles, permissions, identity fields, or the home folder. "
      + "Because this is a field-level atomic patch rather than representation replacement, it does not use If-Match.")
  @RequestBody(description = "An object whose keys are dotted uiPreferences paths",
      required = true, content = @Content(schema = @Schema(type = "object",
      example = "{\"uiPreferences.stylesheet\":\"cedar-dark\",\"uiPreferences.preferredDateFormat\":\"YYYY-MM-DD\"}")))
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The updated user profile",
          content = @Content(schema = @Schema(implementation = CedarUser.class))),
      @ApiResponse(responseCode = "400", description = "A modification is invalid or targets a non-preference field"),
      @ApiResponse(responseCode = "401", description = "Unauthorized"),
      @ApiResponse(responseCode = "403", description = "Another user's profile was targeted"),
      @ApiResponse(responseCode = "404", description = "User not found"),
      @ApiResponse(responseCode = "500", description = "Internal server error")
  })
  public Response updateUser(@Parameter(description = "CEDAR user UUID", required = true)
                             @PathParam(PP_ID) String uuid) throws CedarException {
    CedarRequestContext c = buildRequestContext();

    c.must(c.user()).be(LoggedIn);

    CedarUser currentUser = c.getCedarUser();

    String id = linkedDataUtil.getUserId(uuid);

    if (!id.equals(currentUser.getId())) {
      return CedarResponse.forbidden()
          .id(id)
          .errorKey(CedarErrorKey.UPDATE_OTHER_PROFILE_FORBIDDEN)
          .errorMessage("You are not allowed to update other user's profile!")
          .parameter("currentUserId", currentUser.getId())
          .build();
    }

    JsonNode modifications = c.request().getRequestBody().asJson();

    BackendCallResult<CedarUser> cedarUserBackendCallResult = userService.patchUser(CedarUserId.build(id), modifications);
    if (cedarUserBackendCallResult.isError()) {
      return CedarResponse.from(cedarUserBackendCallResult);
    } else {
      CedarUser updatedUser = cedarUserBackendCallResult.getPayload();

      JsonNode updatedUserNode = JsonMapper.MAPPER.valueToTree(updatedUser);
      // Remove autogenerated _id field to avoid exposing it
      MongoUtils.removeIdField(updatedUserNode);

      return Response.ok().entity(updatedUserNode).build();
    }
  }

  /**
   * Creates a new CEDAR API key for the current user.
   * <p>
   * The optional JSON body may carry a {@code description}; when it is missing or blank a default
   * description carrying today's date is used. A user may hold up to {@link #MAX_API_KEYS} keys.
   * <p>
   * Self-only.
   */
  @POST
  @Timed
  @Path("/{id}/api-keys")
  @Operation(summary = "Create an API key", description = "Atomically appends a newly generated credential to the "
      + "current stored key list. Concurrent creations are serialized and all successful keys survive. Self-only; "
      + "a user may hold at most 20 keys. This command does not use If-Match.")
  @RequestBody(description = "Optional API-key description", required = false,
      content = @Content(schema = @Schema(type = "object", example = "{\"description\":\"analysis pipeline\"}")))
  @ApiResponses({
      @ApiResponse(responseCode = "201", description = "Created. Location names the new key; the body is the "
          + "updated user profile, which is the only place the new key's value is readable",
          content = @Content(schema = @Schema(implementation = CedarUser.class))),
      @ApiResponse(responseCode = "400", description = "The user already has the maximum number of keys"),
      @ApiResponse(responseCode = "401", description = "Unauthorized"),
      @ApiResponse(responseCode = "403", description = "Another user's keys were targeted"),
      @ApiResponse(responseCode = "404", description = "User not found"),
      @ApiResponse(responseCode = "500", description = "Internal server error")
  })
  public Response createApiKey(@Parameter(description = "CEDAR user UUID", required = true)
                               @PathParam(PP_ID) String uuid) throws CedarException {
    CedarRequestContext c = buildRequestContext();
    c.must(c.user()).be(LoggedIn);

    CedarUser currentUser = c.getCedarUser();
    Response forbidden = forbiddenIfNotSelf(uuid, currentUser);
    if (forbidden != null) {
      return forbidden;
    }

    String description = null;
    JsonNode body = c.request().getRequestBody().asJson();
    if (body != null && body.hasNonNull("description")) {
      String candidate = body.get("description").asText().trim();
      if (!candidate.isEmpty()) {
        description = candidate;
      }
    }
    if (description == null) {
      description = "API key created on " + LocalDate.now();
    }

    CedarUserApiKey newApiKey = new CedarUserApiKey();
    newApiKey.setId(UUID.randomUUID().toString());
    newApiKey.setKey(generateRandomApiKey());
    newApiKey.setServiceName("CEDAR");
    newApiKey.setDescription(description);
    newApiKey.setCreationDate(LocalDateTime.now());
    newApiKey.setEnabled(true);

    // The ceiling is checked where the keys are written, not here. Checked against the copy this
    // request holds, it could pass while another request was adding a key of its own.
    BackendCallResult<CedarUser> result =
        userService.addApiKey(CedarUserId.build(currentUser.getId()), newApiKey, MAX_API_KEYS);
    if (result.isError()) {
      return respond(result);
    }
    // Creating a sub-resource answers 201 with a Location, as group, category and artifact creation
    // already do. The body stays the updated profile: the new key's secret value is only ever
    // readable here, so replacing the profile with the key alone would take something away.
    URI keyUri = uriInfo.getAbsolutePathBuilder().path(newApiKey.getId()).build();
    return Response.created(keyUri).entity(profileOf(result)).build();
  }

  /**
   * Regenerates a single API key: the key identified by its non-secret {@code keyId} in the path keeps its
   * description / service name / enabled flag but gets a brand new, random value (and a fresh
   * creation date), so the old value is immediately revoked. The key count is unchanged, so this
   * is always safe - it is the way to "rotate" the last remaining key without dropping below one.
   * <p>
   * Self-only.
   */
  @POST
  @Timed
  @Path("/{id}/api-keys/{keyId}/regenerate")
  @Operation(summary = "Regenerate an API key", description = "Atomically replaces one key's secret and creation "
      + "time while preserving its stable management identifier and metadata. The previous secret is immediately "
      + "revoked. Self-only. This command does not use If-Match.")
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The updated user profile, including the rotated key",
          content = @Content(schema = @Schema(implementation = CedarUser.class))),
      @ApiResponse(responseCode = "401", description = "Unauthorized"),
      @ApiResponse(responseCode = "403", description = "Another user's keys were targeted"),
      @ApiResponse(responseCode = "404", description = "User or key not found"),
      @ApiResponse(responseCode = "500", description = "Internal server error")
  })
  public Response regenerateApiKey(@Parameter(description = "CEDAR user UUID", required = true)
                                   @PathParam(PP_ID) String uuid,
                                   @Parameter(description = "Stable, non-secret API-key management identifier", required = true)
                                   @PathParam("keyId") String keyId)
      throws CedarException {
    CedarRequestContext c = buildRequestContext();
    c.must(c.user()).be(LoggedIn);

    CedarUser currentUser = c.getCedarUser();
    Response forbidden = forbiddenIfNotSelf(uuid, currentUser);
    if (forbidden != null) {
      return forbidden;
    }

    return respond(userService.regenerateApiKey(CedarUserId.build(currentUser.getId()), keyId,
        generateRandomApiKey(), LocalDateTime.now()));
  }

  /**
   * Deletes a single API key. Refused when it would leave the user with no enabled key: a user must
   * always keep at least one active key (regenerate it instead of deleting the last one). Deleting a
   * disabled key is always allowed, since a disabled key is not the account's working one.
   * <p>
   * Self-only.
   */
  @DELETE
  @Timed
  @Path("/{id}/api-keys/{keyId}")
  @Operation(summary = "Delete an API key", description = "Atomically removes one key from the current stored key "
      + "list. Deleting the last enabled key is refused so the account retains a working credential. Self-only. "
      + "This command does not use If-Match.")
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The updated user profile",
          content = @Content(schema = @Schema(implementation = CedarUser.class))),
      @ApiResponse(responseCode = "400", description = "Deleting the key would leave no enabled credential"),
      @ApiResponse(responseCode = "401", description = "Unauthorized"),
      @ApiResponse(responseCode = "403", description = "Another user's keys were targeted"),
      @ApiResponse(responseCode = "404", description = "User or key not found"),
      @ApiResponse(responseCode = "500", description = "Internal server error")
  })
  public Response deleteApiKey(@Parameter(description = "CEDAR user UUID", required = true)
                               @PathParam(PP_ID) String uuid,
                               @Parameter(description = "Stable, non-secret API-key management identifier", required = true)
                               @PathParam("keyId") String keyId)
      throws CedarException {
    CedarRequestContext c = buildRequestContext();
    c.must(c.user()).be(LoggedIn);

    CedarUser currentUser = c.getCedarUser();
    Response forbidden = forbiddenIfNotSelf(uuid, currentUser);
    if (forbidden != null) {
      return forbidden;
    }

    // Both the lookup of the key and the "keep one active key" rule are applied where the keys are
    // written, against the stored list rather than against the copy this request holds.
    return respond(userService.deleteApiKey(CedarUserId.build(currentUser.getId()), keyId));
  }

  /**
   * Returns a 403 response when {@code uuid} does not identify the current user; otherwise null.
   */
  private Response forbiddenIfNotSelf(String uuid, CedarUser currentUser) {
    String id = linkedDataUtil.getUserId(uuid);
    if (!id.equals(currentUser.getId())) {
      return CedarResponse.forbidden()
          .id(id)
          .errorKey(CedarErrorKey.UPDATE_OTHER_PROFILE_FORBIDDEN)
          .errorMessage("You are not allowed to manage another user's API keys!")
          .parameter("currentUserId", currentUser.getId())
          .build();
    }
    return null;
  }

  /**
   * Renders the outcome of a key change as the updated user JSON, mirroring the shape of
   * GET /users/{id}. A refusal carries its own status and message from the layer that applied it.
   */
  private Response respond(BackendCallResult<CedarUser> callResult) {
    if (callResult.isError()) {
      return CedarResponse.from(callResult);
    }
    return Response.ok().entity(profileOf(callResult)).build();
  }

  /** The user profile as a client sees it, without the autogenerated Mongo id. */
  private JsonNode profileOf(BackendCallResult<CedarUser> callResult) {
    JsonNode updatedUserNode = JsonMapper.MAPPER.valueToTree(callResult.getPayload());
    MongoUtils.removeIdField(updatedUserNode);
    return updatedUserNode;
  }

  /**
   * Generates a new, cryptographically-random API key rendered as 64 lowercase hex characters
   * (256 bits of entropy), matching the format of the keys CEDAR generates at user provisioning time.
   */
  private String generateRandomApiKey() {
    SecureRandom random = new SecureRandom();
    byte[] bytes = new byte[32];
    random.nextBytes(bytes);
    StringBuilder sb = new StringBuilder(64);
    for (byte b : bytes) {
      sb.append(String.format("%02x", b));
    }
    return sb.toString();
  }
}